Updates to pathfinding

+++ App_Code/tests/calculatorTests.cs
@@ -1,7 +1,6 @@
 ﻿using System;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
-using Calculator_ns;
-using Point_ns;
+using PathFinding;
 
 namespace calcTest
 {
@@ -11,10 +10,10 @@ namespace calcTest
         [TestMethod]
         public void ifStartEndPointSameThenScaleIsZero()
         {
-            Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(0.0, 0.0), new Point(0.0, 0.0), 20),
+            Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(0, 0), new Point(0, 0), 20),
             "The scale of zero is correctely calculated");
 
-            Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(5.14, 3.11), new Point(5.14, 3.11), 11.3),
+            Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point((float)5.14, (float)3.11), new Point((float)5.14, (float)3.11), 11.3),
             "The scale of zero is correctely calculated");
 
         }
@@ -22,10 +21,10 @@ namespace calcTest
         [TestMethod]
         public void ifLengthIsZeroThenScaleIsZero()
         {
-             Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(0.0, 0.0), new Point(0.0, 0.0), 0),
+             Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(0, 0), new Point(0, 0), 0),
             "The scale of zero is correctely calculated");
 
-             Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point(3.25, 6.6666), new Point(8.99, 0.0005), 0),
+             Assert.AreEqual<double>(0, CoordinateCalculator.getScale(new Point((float)3.25, (float)6.6666), new Point((float)8.99, (float)0.0005), 0),
             "The scale of zero is correctely calculated");     
 
         }
@@ -33,10 +32,10 @@ namespace calcTest
         [TestMethod]
         public void ifDistanceIsVerticalOrHorizontal()
         {
-            Assert.AreEqual<double>(0.5, CoordinateCalculator.getScale(new Point(0.0, 0.0), new Point(1.0, 0.0), 2),
+            Assert.AreEqual<double>(0.5, CoordinateCalculator.getScale(new Point(0, 0), new Point(1, 0), 2),
             "The scale of 1/2 is correctely calculated");
             
-            Assert.AreEqual<double>(0.5, CoordinateCalculator.getScale(new Point(0.0, 0.0), new Point(0.0, 1.0), 2),
+            Assert.AreEqual<double>(0.5, CoordinateCalculator.getScale(new Point(0, 0), new Point(0, 1), 2),
             "The scale of 1/2 is correctely calculated");
 
         }
@@ -45,7 +44,7 @@ namespace calcTest
         [TestMethod]
         public void testDistanceCalculation()
         {
-            Assert.AreEqual<double>(5, CoordinateCalculator.euclideanDistance(new Point(1.0, 2.0), new Point(4.0, 6.0)),
+            Assert.AreEqual<double>(5, CoordinateCalculator.euclideanDistance(new Point(1, 2), new Point(4, 6)),
             "the distance of 5 between points (1,2), and (4,6) is calculated correctely");
         }
 
@@ -53,13 +52,12 @@ namespace calcTest
         [TestMethod]
         public void testScaleCalculation()
         {
-            Assert.AreEqual<double>(2.5, CoordinateCalculator.getScale(new Point(0.0, 0.0), new Point(3.0, 4.0), 2),
+            Assert.AreEqual<double>(2.5, CoordinateCalculator.getScale(new Point(0, 0), new Point(3, 4), 2),
             "The scale of 5/2 is correctely calculated");
         }
 
         
-       
-
+     
        
     }
 }
